separated bishop.cpp into a header and the corresponding file.
